Dev vercel pg integration

+++ app/api/tasks/[id]/route.ts
@@ -1,6 +1,7 @@
 import { query } from '../../../../lib/db';
 import { getServerSession } from 'next-auth';
 import { authOptions, getUserID } from '@/lib/auth';
+import { sql } from '@vercel/postgres';
 
 export async function GET(req: Request, { params }: { params: { id: string } }) {
     const session = await getServerSession(authOptions);
@@ -12,9 +13,15 @@ export async function GET(req: Request, { params }: { params: { id: string } }) 
     const { id } = params;
 
     try {
-        const { rows } = await query('SELECT * FROM tasks WHERE id = $1', [id]);
-        const task = rows[0];
-        if (task.userId !== sessionUserId) {
+        let task;
+        if (process.env.NODE_ENV === 'production') {
+            const { rows } = await sql`SELECT * FROM tasks WHERE id = ${id} AND user_id = ${sessionUserId}`;
+            task = rows[0];
+        } else {
+            const { rows } = await query('SELECT * FROM tasks WHERE id = $1 AND user_id = $2', [id, sessionUserId]);
+            task = rows[0];
+        }
+        if (!task) {
             return new Response(JSON.stringify({ error: 'Forbidden: user id does not match session user id' }), { status: 403 });
         }
         return new Response(JSON.stringify(task), { status: 200 });
@@ -35,22 +42,21 @@ export async function PUT(req: Request, { params }: { params: { id: string } }) 
     const { task_name, deadline, total_set, current_set, is_complete } = await req.json();
 
     try {
-        const { rows } = await query('SELECT user_id FROM tasks WHERE id = $1', [id]);
-        const user_id = rows[0].user_id;
-        if (user_id !== session_user_id) {
+        var task;
+        if (process.env.NODE_ENV === 'production') {
+            const { rows } = await sql`UPDATE tasks SET task_name = ${task_name}, deadline = ${deadline}, total_set = ${total_set}, current_set = ${current_set}, is_complete = ${is_complete} WHERE id = ${id} AND user_id = ${session_user_id} RETURNING *`;
+            task = rows[0];
+        } else {
+            const { rows } = await query(
+                'UPDATE tasks SET task_name = $1, deadline = $2, total_set = $3, current_set = $4, is_complete = $5 WHERE id = $6 AND user_id = $7 RETURNING *',
+                [task_name, deadline, total_set, current_set, is_complete, id, session_user_id]
+            );
+            task = rows[0];
+        }
+        if (!task) {
             return new Response(JSON.stringify({ error: 'Forbidden: user_id in updating task does not match session user id' }), { status: 403 });
         }
-    } catch (error) {
-        console.error('Database query failed:', error);
-        return new Response(JSON.stringify({ error: 'Failed to update task' }), { status: 500 });
-    }
-
-    try {
-        const { rows } = await query(
-            'UPDATE tasks SET task_name = $1, deadline = $2, total_set = $3, current_set = $4, is_complete = $5 WHERE id = $6 RETURNING *',
-            [task_name, deadline, total_set, current_set, is_complete, id]
-        );
-        return new Response(JSON.stringify(rows[0]), { status: 200 });
+        return new Response(JSON.stringify(task), { status: 200 });
     } catch (error) {
         console.error('Database query failed:', error);
         return new Response(JSON.stringify({ error: 'Failed to update task' }), { status: 500 });
@@ -67,13 +73,11 @@ export async function DELETE(req: Request, { params }: { params: { id: string } 
     const { id } = params;
 
     try {
-        const { rows } = await query('SELECT user_id FROM tasks WHERE id = $1', [id]);
-        const user_id = rows[0].user_id;
-        if (user_id !== session_user_id) {
-            return new Response(JSON.stringify({ error: 'Forbidden: user id does not match session user id' }), { status: 403 });
+        if (process.env.NODE_ENV === 'production') {
+            await sql`DELETE FROM tasks WHERE id = ${id} AND user_id = ${session_user_id}`;
+        } else {
+            await query('DELETE FROM tasks WHERE id = $1 AND user_id = $2', [id, session_user_id]);
         }
-
-        await query('DELETE FROM tasks WHERE id = $1', [id]);
         return new Response(null, { status: 204 });
     } catch (error) {
         console.error('Database query failed:', error);
